requirements.txt
pyyaml==3.11
autobahn==0.14.1